tot actualitzat, fa coses raresgit add .

+++ wp-content/plugins/contact-form-7/includes/functions.php
@@ -7,7 +7,8 @@ function wpcf7_plugin_path( $path = '' ) {
 function wpcf7_plugin_url( $path = '' ) {
 	$url = plugins_url( $path, WPCF7_PLUGIN );
 
-	if ( is_ssl() && 'http:' == substr( $url, 0, 5 ) ) {
+	if ( is_ssl()
+	and 'http:' == substr( $url, 0, 5 ) ) {
 		$url = 'https:' . substr( $url, 5 );
 	}
 
@@ -37,32 +38,6 @@ function wpcf7_verify_nonce( $nonce, $action = 'wp_rest' ) {
 
 function wpcf7_create_nonce( $action = 'wp_rest' ) {
 	return wp_create_nonce( $action );
-}
-
-function wpcf7_blacklist_check( $target ) {
-	$mod_keys = trim( get_option( 'blacklist_keys' ) );
-
-	if ( empty( $mod_keys ) ) {
-		return false;
-	}
-
-	$words = explode( "\n", $mod_keys );
-
-	foreach ( (array) $words as $word ) {
-		$word = trim( $word );
-
-		if ( empty( $word ) || 256 < strlen( $word ) ) {
-			continue;
-		}
-
-		$pattern = sprintf( '#%s#i', preg_quote( $word, '#' ) );
-
-		if ( preg_match( $pattern, $target ) ) {
-			return true;
-		}
-	}
-
-	return false;
 }
 
 function wpcf7_array_flatten( $input ) {
@@ -106,6 +81,10 @@ function wpcf7_use_really_simple_captcha() {
 function wpcf7_validate_configuration() {
 	return apply_filters( 'wpcf7_validate_configuration',
 		WPCF7_VALIDATE_CONFIGURATION );
+}
+
+function wpcf7_autop_or_not() {
+	return (bool) apply_filters( 'wpcf7_autop_or_not', WPCF7_AUTOP );
 }
 
 function wpcf7_load_js() {
@@ -188,7 +167,8 @@ function wpcf7_register_post_types() {
 function wpcf7_version( $args = '' ) {
 	$defaults = array(
 		'limit' => -1,
-		'only_major' => false );
+		'only_major' => false,
+	);
 
 	$args = wp_parse_args( $args, $defaults );
 
@@ -248,26 +228,33 @@ function wpcf7_enctype_value( $enctype ) {
 
 function wpcf7_rmdir_p( $dir ) {
 	if ( is_file( $dir ) ) {
-		if ( ! $result = @unlink( $dir ) ) {
-			$stat = @stat( $dir );
-			$perms = $stat['mode'];
-			@chmod( $dir, $perms | 0200 ); // add write for owner
+		$file = $dir;
 
-			if ( ! $result = @unlink( $dir ) ) {
-				@chmod( $dir, $perms );
-			}
+		if ( @unlink( $file ) ) {
+			return true;
 		}
 
-		return $result;
+		$stat = stat( $file );
+
+		if ( @chmod( $file, $stat['mode'] | 0200 ) ) { // add write for owner
+			if ( @unlink( $file ) ) {
+				return true;
+			}
+
+			@chmod( $file, $stat['mode'] );
+		}
+
+		return false;
 	}
 
 	if ( ! is_dir( $dir ) ) {
 		return false;
 	}
 
-	if ( $handle = @opendir( $dir ) ) {
+	if ( $handle = opendir( $dir ) ) {
 		while ( false !== ( $file = readdir( $handle ) ) ) {
-			if ( $file == "." || $file == ".." ) {
+			if ( $file == "."
+			or $file == ".." ) {
 				continue;
 			}
 
@@ -277,7 +264,12 @@ function wpcf7_rmdir_p( $dir ) {
 		closedir( $handle );
 	}
 
-	return @rmdir( $dir );
+	if ( false !== ( $files = scandir( $dir ) )
+	and ! array_diff( $files, array( '.', '..' ) ) ) {
+		return rmdir( $dir );
+	}
+
+	return false;
 }
 
 /* From _http_build_query in wp-includes/functions.php */
@@ -298,7 +290,7 @@ function wpcf7_build_query( $args, $key = '' ) {
 			$v = '0';
 		}
 
-		if ( is_array( $v ) || is_object( $v ) ) {
+		if ( is_array( $v ) or is_object( $v ) ) {
 			array_push( $ret, wpcf7_build_query( $v, $k ) );
 		} else {
 			array_push( $ret, $k . '=' . urlencode( $v ) );
@@ -348,13 +340,127 @@ function wpcf7_is_localhost() {
 }
 
 function wpcf7_deprecated_function( $function, $version, $replacement ) {
-	$trigger_error = apply_filters( 'deprecated_function_trigger_error', true );
-
-	if ( WP_DEBUG && $trigger_error ) {
+	if ( WP_DEBUG ) {
 		if ( function_exists( '__' ) ) {
-			trigger_error( sprintf( __( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', 'contact-form-7' ), $function, $version, $replacement ) );
+			trigger_error(
+				sprintf(
+					/* translators: 1: PHP function name, 2: version number, 3: alternative function name */
+					__( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', 'contact-form-7' ),
+					$function, $version, $replacement
+				),
+				E_USER_DEPRECATED
+			);
 		} else {
-			trigger_error( sprintf( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', $function, $version, $replacement ) );
+			trigger_error(
+				sprintf(
+					'%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.',
+					$function, $version, $replacement
+				),
+				E_USER_DEPRECATED
+			);
 		}
 	}
+}
+
+function wpcf7_apply_filters_deprecated( $tag, $args, $version, $replacement ) {
+	if ( ! has_filter( $tag ) ) {
+		return $args[0];
+	}
+
+	if ( WP_DEBUG ) {
+		trigger_error(
+			sprintf(
+				/* translators: 1: WordPress hook name, 2: version number, 3: alternative hook name */
+				__( '%1$s is <strong>deprecated</strong> since Contact Form 7 version %2$s! Use %3$s instead.', 'contact-form-7' ),
+				$tag, $version, $replacement
+			),
+			E_USER_DEPRECATED
+		);
+	}
+
+	return apply_filters_ref_array( $tag, $args );
+}
+
+function wpcf7_doing_it_wrong( $function, $message, $version ) {
+	if ( WP_DEBUG ) {
+		if ( function_exists( '__' ) ) {
+			if ( $version ) {
+				$version = sprintf(
+					/* translators: %s: Contact Form 7 version number. */
+					__( '(This message was added in Contact Form 7 version %s.)', 'contact-form-7' ),
+					$version
+				);
+			}
+
+			trigger_error(
+				sprintf(
+					/* translators: Developer debugging message. 1: PHP function name, 2: Explanatory message, 3: Contact Form 7 version number. */
+					__( '%1$s was called incorrectly. %2$s %3$s', 'contact-form-7' ),
+					$function,
+					$message,
+					$version
+				),
+				E_USER_NOTICE
+			);
+		} else {
+			if ( $version ) {
+				$version = sprintf(
+					'(This message was added in Contact Form 7 version %s.)',
+					$version
+				);
+			}
+
+			trigger_error(
+				sprintf(
+					'%1$s was called incorrectly. %2$s %3$s',
+					$function,
+					$message,
+					$version
+				),
+				E_USER_NOTICE
+			);
+		}
+	}
+}
+
+function wpcf7_log_remote_request( $url, $request, $response ) {
+	$log = sprintf(
+		/* translators: 1: response code, 2: message, 3: body, 4: URL */
+		__( 'HTTP Response: %1$s %2$s %3$s from %4$s', 'contact-form-7' ),
+		(int) wp_remote_retrieve_response_code( $response ),
+		wp_remote_retrieve_response_message( $response ),
+		wp_remote_retrieve_body( $response ),
+		$url
+	);
+
+	$log = apply_filters( 'wpcf7_log_remote_request',
+		$log, $url, $request, $response
+	);
+
+	if ( $log ) {
+		trigger_error( $log );
+	}
+}
+
+function wpcf7_anonymize_ip_addr( $ip_addr ) {
+	if ( ! function_exists( 'inet_ntop' )
+	or ! function_exists( 'inet_pton' ) ) {
+		return $ip_addr;
+	}
+
+	$packed = inet_pton( $ip_addr );
+
+	if ( false === $packed ) {
+		return $ip_addr;
+	}
+
+	if ( 4 == strlen( $packed ) ) { // IPv4
+		$mask = '255.255.255.0';
+	} elseif ( 16 == strlen( $packed ) ) { // IPv6
+		$mask = 'ffff:ffff:ffff:0000:0000:0000:0000:0000';
+	} else {
+		return $ip_addr;
+	}
+
+	return inet_ntop( $packed & inet_pton( $mask ) );
 }
